Move edit view action to view side menu component
