+++ src/test/java/school/redrover/page/view/FolderViewPage.java
@@ -5,8 +5,7 @@ import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
 import org.openqa.selenium.support.ui.ExpectedConditions;
 import school.redrover.page.common.BasePage;
-import school.redrover.page.components.BaseSideMenuComponent;
-import school.redrover.page.view.config.ViewMyConfigPage;
+import school.redrover.page.components.ViewSideMenuComponent;
 
 public class FolderViewPage extends BasePage {
 
@@ -28,24 +27,21 @@ public class FolderViewPage extends BasePage {
     @FindBy(name = "Submit")
     private WebElement buttonSave;
 
-    @FindBy(xpath = "//a[contains(@href, '/configure') and .//span[normalize-space()='Edit View']]")
-    private WebElement editViewButton;
-
     @FindBy(id = "description-link")
     private WebElement editDescription;
 
     @FindBy(css = "#description-edit-form .description-cancel-button")
     private WebElement buttonCancel;
 
-    private final BaseSideMenuComponent<FolderViewPage> baseSideMenu;
+    private final ViewSideMenuComponent<FolderViewPage> viewSideMenu;
 
     public FolderViewPage(WebDriver driver) {
         super(driver);
-        this.baseSideMenu = new BaseSideMenuComponent<>(getDriver(), this);
+        this.viewSideMenu = new ViewSideMenuComponent<>(getDriver(), this);
     }
 
-    public BaseSideMenuComponent<FolderViewPage> getSideMenu() {
-        return this.baseSideMenu;
+    public ViewSideMenuComponent<FolderViewPage> getSideMenu() {
+        return this.viewSideMenu;
     }
 
     public FolderViewPage addDescription(String expectedDescription) {
@@ -67,11 +63,6 @@ public class FolderViewPage extends BasePage {
 
     public String getDescriptionText() {
         return getWait10().until(ExpectedConditions.refreshed(ExpectedConditions.visibilityOf(content))).getText();
-    }
-
-    public ViewMyConfigPage clickEditView() {
-        getWait5().until(ExpectedConditions.elementToBeClickable(editViewButton)).click();
-        return new ViewMyConfigPage(getDriver());
     }
 
     public FolderViewPage editDescription(String newDescriptionText) {
